fix: resolve refetch when click event is triggered
