+++ src/containers/Tables/BidTable/BidTableCard/BidTableCard.tsx
@@ -1,15 +1,18 @@
 //- React Imports
-import { memo } from 'react';
+import { memo, useState } from 'react';
 
 //- Components Imports
-import { Artwork, OptionDropdown } from 'components';
+import { Artwork, OptionDropdown, Overlay } from 'components';
 import { Option } from 'components/Dropdowns/OptionDropdown/OptionDropdown';
 
 //- Library Imports
 import { ethers } from 'ethers';
 
 //- Constants Imports
-import { AltText, Labels, TestId } from '../BidTable.constants';
+import { AltText, Labels, Modal, TestId } from '../BidTable.constants';
+
+//- Container Imports
+import { CancelBid, MakeABid } from 'containers';
 
 //- Types Imports
 import { Actions, ActionKeys, BidTableData } from '../BidTable.types';
@@ -26,63 +29,101 @@ import moreIcon from 'assets/more-vertical.svg';
 
 const BidTableCard = (props: any) => {
 	const dimensions = usePageWidth();
+	const [modal, setModal] = useState<Modal | undefined>();
+	const [selectedBid, setSelectedBid] = useState<BidTableData>();
 	const bid: BidTableData = props.data;
+
+	const openMakeBid = (bid: BidTableData) => {
+		setSelectedBid(bid);
+		setModal(Modal.MAKE_A_BID);
+	};
+
+	const openCancelBid = (bid: BidTableData) => {
+		setSelectedBid(bid);
+		setModal(Modal.CANCEL);
+	};
+
+	const closeModal = () => {
+		setModal(undefined);
+		setSelectedBid(undefined);
+	};
 
 	const onSelectOption = (option: Option) => {
 		switch (option.title) {
 			case ActionKeys.REBID:
-				props.openMakeBid?.(bid);
+				openMakeBid?.(bid);
 				return;
 			case ActionKeys.CANCEL_BID:
-				props.openCancelBid?.(bid);
+				openCancelBid?.(bid);
 				return;
 		}
 	};
 
 	return (
-		<div className={styles.CardContainer} data-testid={TestId.CARD_CONTAINER}>
-			<div className={styles.Content}>
-				<div className={styles.DomainInfoContainer}>
-					<Artwork
-						subtext={'0://' + bid.domainName}
-						disableInteraction
-						metadataUrl={bid.domainMetadataUrl}
-						id={bid.domainId}
-						style={{ maxWidth: handleDomainNameWidth(dimensions.pageWidth) }}
+		<>
+			{modal === Modal.MAKE_A_BID && selectedBid && (
+				<MakeABid
+					domain={selectedBid.domain}
+					onBid={props.refetch}
+					onClose={() => setModal(undefined)}
+					paymentTokenInfo={selectedBid.paymentTokenInfo}
+				/>
+			)}
+			{modal === Modal.CANCEL && selectedBid && (
+				<Overlay onClose={() => setModal(undefined)} open>
+					<CancelBid
+						bidNonce={selectedBid.bidNonce}
+						domainId={selectedBid.domainId}
+						onSuccess={props.refetch}
+						onClose={closeModal}
+						paymentTokenInfo={selectedBid.paymentTokenInfo}
 					/>
+				</Overlay>
+			)}
+			<div className={styles.CardContainer} data-testid={TestId.CARD_CONTAINER}>
+				<div className={styles.Content}>
+					<div className={styles.DomainInfoContainer}>
+						<Artwork
+							subtext={'0://' + bid.domainName}
+							disableInteraction
+							metadataUrl={bid.domainMetadataUrl}
+							id={bid.domainId}
+							style={{ maxWidth: handleDomainNameWidth(dimensions.pageWidth) }}
+						/>
 
-					<div className={styles.BidContainer}>
-						<div className={styles.TextContainer}>
-							<div className={styles.Title}>{Labels.YOUR_BID}</div>
-							<div className={styles.Bid} data-testid={TestId.YOUR_BID}>
-								{ethers.utils.formatEther(bid.yourBid.toString())}{' '}
-								{bid.paymentTokenInfo.symbol}
+						<div className={styles.BidContainer}>
+							<div className={styles.TextContainer}>
+								<div className={styles.Title}>{Labels.YOUR_BID}</div>
+								<div className={styles.Bid} data-testid={TestId.YOUR_BID}>
+									{ethers.utils.formatEther(bid.yourBid.toString())}{' '}
+									{bid.paymentTokenInfo.symbol}
+								</div>
 							</div>
-						</div>
 
-						<div className={styles.TextContainer}>
-							<div className={styles.Title}>{Labels.TOP_BID}</div>
-							<div className={styles.Bid} data-testid={TestId.HIGHEST_BID}>
-								{ethers.utils.formatEther(bid.highestBid)}{' '}
-								{bid.paymentTokenInfo.symbol}
+							<div className={styles.TextContainer}>
+								<div className={styles.Title}>{Labels.TOP_BID}</div>
+								<div className={styles.Bid} data-testid={TestId.HIGHEST_BID}>
+									{ethers.utils.formatEther(bid.highestBid)}{' '}
+									{bid.paymentTokenInfo.symbol}
+								</div>
 							</div>
 						</div>
 					</div>
-				</div>
-				<div className={styles.OptionsContainer}>
-					<OptionDropdown
-						className={styles.MoreDropdown}
-						onSelect={onSelectOption}
-						options={Actions}
-						data-testid={TestId.OPTIONS}
-					>
-						<button className={styles.Button}>
-							<img alt={AltText.MORE_OPTIONS} src={moreIcon} />
-						</button>
-					</OptionDropdown>
+					<div className={styles.OptionsContainer}>
+						<OptionDropdown
+							className={styles.MoreDropdown}
+							onSelect={onSelectOption}
+							options={Actions}
+							data-testid={TestId.OPTIONS}
+						>
+							<button className={styles.Button}>
+								<img alt={AltText.MORE_OPTIONS} src={moreIcon} />
+							</button>
+						</OptionDropdown>
+					</div>
 				</div>
 			</div>
-		</div>
+		</>
 	);
 };
