+++ src/containers/flows/MakeABid/MakeABid.tsx
@@ -196,8 +196,6 @@ const MakeABid = ({
 			);
 			setIsBidPlaced(true);
 			setStepContent(StepContent.Success);
-			// refetch bid data once bid successful
-			onBid();
 		} catch (e) {
 			if (!isMounted.current) {
 				return;
@@ -221,6 +219,11 @@ const MakeABid = ({
 	const onStepNavigation = (i: number) => {
 		setCurrentStep(i);
 		setStepContent(i);
+	};
+
+	const handleClose = () => {
+		onBid();
+		onClose();
 	};
 
 	/**
@@ -340,7 +343,7 @@ const MakeABid = ({
 				tokenBalance={tokenBalance}
 				highestBid={bidData?.highestBid?.amount}
 				bid={bid}
-				onClose={onClose}
+				onClose={handleClose}
 				paymentTokenInfo={paymentTokenInfo}
 			/>
 		),
